requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.2
requests==2.26.0
apscheduler==3.8.1
